Refactor Deployment
